requirements.txt
flask>=2.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.11.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
reportlab>=4.0.0
python-docx>=1.1.0
nltk>=3.8.0
networkx>=3.1
